moved process_and_get_fields2expt_type_obj_dict and associated functions from ExperimentChecklists2Json.py to ExperimentType.py
